Revisions to collaborators

+++ app/controllers/collaborators_controller.rb
@@ -2,6 +2,7 @@ class CollaboratorsController < ApplicationController
 
   def create
     @collaborator = Collaborator.new(collaborator_params)
+    authorize @collaborator
     @collaborator.save
 
     if @collaborator.save
@@ -14,14 +15,15 @@ class CollaboratorsController < ApplicationController
   end
 
   def destroy
-    @collaborator = Collaborator.find_by_user_id_and_wiki_id(params[:user_id],params[:wiki_id])
+    @collaborator = Collaborator.find_by(user_id: params[:user_id], wiki_id: params[:wiki_id])
+    authorize @collaborator
 
     if @collaborator.destroy
       flash[:notice] = "Collaborator removed."
     else
       flash[:alert] = "Collaborator removal failed."
     end
-    
+
     redirect_back(fallback_location: root_path)
   end
 
@@ -29,10 +31,6 @@ class CollaboratorsController < ApplicationController
 
   def collaborator_params
     params.permit(:user_id, :wiki_id)
-  end
-
-  def collaborator_users(wiki_id)
-    Collaborator.where(wiki_id: wiki_id).collect(&:user)
   end
 
 end
